Remove Network.TLS.Handshake.QUIC
